HeartBeat is saved for each user and Average value is shown on the screen

+++ app/src/main/java/com/example/taskbeat/data/HealthDao.kt
@@ -1,10 +1,6 @@
 package com.example.taskbeat.data
 
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import androidx.room.Update
+import androidx.room.*
 import com.example.taskbeat.model.Health
 import kotlinx.coroutines.flow.Flow
 
@@ -15,13 +11,13 @@ interface HealthDao {
     suspend fun insertHealthData(health: Health)
 
     @Query("SELECT * FROM health WHERE user_id = :userId")
-    fun getHealthDataByUserId(userId: Int): Flow<Health?>
+    fun getHealthDataByUserId(userId: Long): Flow<Health?>
 
     @Update
     suspend fun updateHealthData(health: Health)
 
     @Query("DELETE FROM health WHERE user_id = :userId")
-    suspend fun deleteHealthDataByUserId(userId: Int)
+    suspend fun deleteHealthDataByUserId(userId: Long)
 
     @Query("SELECT * FROM health")
     fun getAllHealthData(): Flow<List<Health>>
